feat(nixos): don't kill ZFS if passphrase is incorrect
Killing ZFS is necessary to end the prompt shown on the console.
However, if the entered passhrase was not correct, we won't be able to
try again.

+++ features/nixos/remote-unlock.nix
@@ -45,8 +45,11 @@ in {
           cat <<EOF > /root/.profile
           if pgrep -x "zfs" > /dev/null
           then
-            zfs load-key -a
-            killall zfs
+            if zfs load-key -a; then
+              killall zfs
+            else
+              echo "zfs passphrase invalid. Please try again."
+            fi
           else
             echo "zfs not running -- maybe the pool is taking some time to load for some unforseen reason."
           fi
